server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pydantic==2.3.0
numpy>=1.21.0,<2.0.0
h5py>=3.1.0
tensorflow>=2.10.0,<2.16.0
mtcnn>=0.1.1
speechbrain
torch
sounddevice
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-learn>=1.0.0
librosa>=0.8.0
waitress>=2.1.0
